Refactor All Handlers Together
During development we had some handlers that were done during setup,
much better to have them all together.
