+++ packages/ilios-common/addon-test-support/msw/handlers/get-auth-logout.js
@@ -1,0 +1,6 @@
+import { http, HttpResponse } from 'msw';
+
+// GET /auth/logout
+export default http.get('/auth/logout', () => {
+  return new HttpResponse(null, { status: 200 });
+});
