+++ packages/ilios-common/addon-test-support/msw/handlers/get-auth-whoami.js
@@ -1,0 +1,8 @@
+import { http, HttpResponse } from 'msw';
+
+// GET /auth/whoami
+export default http.get('/auth/whoami', () => {
+  return HttpResponse.json({
+    userId: 4136,
+  });
+});
