+++ packages/ilios-common/addon-test-support/msw/handlers/get-errors.js
@@ -1,0 +1,6 @@
+import { http, HttpResponse } from 'msw';
+
+// POST /errors - swallow errors
+export default http.post('/errors', () => {
+  return new HttpResponse(null, { status: 204 });
+});
