+++ packages/ilios-common/addon-test-support/msw/handlers/index.js
@@ -1,3 +1,10 @@
+import postAuthLogin from './post-auth-login';
+import getAuthLogout from './get-auth-logout';
+import getAuthWhoami from './get-auth-whoami';
+import getAuthToken from './get-auth-token';
+import getApplicationConfig from './get-application-config';
+import getErrors from './get-errors';
+
 import getCohorts from './get-cohorts';
 import getCourses from './get-courses';
 import getPendingUserUpdates from './get-pending-user-updates';
@@ -7,17 +14,26 @@ import getUserEvents from './get-user-events';
 import postProgramYear from './post-program-year';
 import postReports from './post-reports';
 import postUpload from './post-upload';
+
 import { genericHandlers } from './generic-crud';
 
-export default [
-  getCohorts,
-  getCourses,
-  getPendingUserUpdates,
-  getSchoolEvents,
-  getSessions,
-  getUserEvents,
-  postProgramYear,
-  postReports,
-  postUpload,
-  ...genericHandlers,
-];
+export function getHandlers(config) {
+  return [
+    postAuthLogin,
+    getAuthLogout,
+    getAuthWhoami,
+    getAuthToken,
+    getErrors,
+    getApplicationConfig(config),
+    getCohorts,
+    getCourses,
+    getPendingUserUpdates,
+    getSchoolEvents,
+    getSessions,
+    getUserEvents,
+    postProgramYear,
+    postReports,
+    postUpload,
+    ...genericHandlers,
+  ];
+}
